reorganized settings vc
